end of part4

+++ part4/blogs/controllers/blogs.js
@@ -1,15 +1,34 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 require('express-async-errors')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const blog = await new Blog(request.body).save()
-    response.status(201).json(blog).end()
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    } 
+    const user = await User.findById(decodedToken.id)
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    })
+    const newBlog = await blog.save()
+    user.blogs = user.blogs.concat(newBlog._id)
+    await user.save()
+    response.status(201).json(newBlog).end()
 })
 
 
@@ -19,8 +38,20 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    } 
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if ( blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }
+    response.status(401).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -33,7 +64,7 @@ blogsRouter.put('/:id', async (request, response) => {
         likes: body.likes,
     }
 
-    const newBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
+    const newBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response.json(newBlog)
 })
